switch to making the rpm contain the unpacked .iso contents
This *might* make the appliance a fraction more efficient in terms of
storage (no compressed image nested in another compressed image) and
performance (no need to go through the loopback device).

It will also allow us to modify the repository, which seems to be
required since at this stage of the image build.  This is because the
.iso provided doesn't have a proper signature in content.key, because
the BS bs_signer backend service run asynchronously so hasn't yet signed
the .iso.  Consequently we'll have to ensure a valid signature
ourselves.