add a fake product key
At the kiwi_post_run stage of the image build, the .iso provided doesn't
yet have a proper signature in content.key, because the BS bs_signer
backend service run asynchronously so hasn't yet signed the .iso.
Consequently we'll have to ensure a valid signature ourselves.  For now
we'll just disable key-checking, so it's good enough to have any key as
long as it's the right format to keep "zypper refresh" happy.